Reconfigure taco-auth providers folder to have an eip4361 folder and common elements for both ei4361 provider and single sign-on provider. Move user address context variable for external eip4361 to taco-auth as well. Update relevant imports.
