+++ packages/taco-auth/src/auth-sig.ts
@@ -1,8 +1,10 @@
 import { EthAddressSchema } from '@nucypher/shared';
 import { z } from 'zod';
 
-import { EIP4361_AUTH_METHOD } from './auth-provider';
-import { EIP4361TypedDataSchema } from './providers';
+import {
+  EIP4361_AUTH_METHOD,
+  EIP4361TypedDataSchema,
+} from './providers/eip4361/common';
 
 export const authSignatureSchema = z.object({
   signature: z.string(),
