+++ packages/taco-auth/src/providers/eip4361/external-eip4361.ts
@@ -1,7 +1,11 @@
 import { SiweMessage } from 'siwe';
 
-import { EIP4361_AUTH_METHOD } from '../auth-provider';
-import { AuthSignature } from '../auth-sig';
+import { AuthSignature } from '../../auth-sig';
+
+import { EIP4361_AUTH_METHOD } from './common';
+
+export const USER_ADDRESS_PARAM_EXTERNAL_EIP4361 =
+  ':userAddressExternalEIP4361';
 
 export class SingleSignOnEIP4361AuthProvider {
   public static async fromExistingSiweInfo(
