+++ packages/taco-auth/src/providers/index.ts
@@ -1,2 +1,2 @@
-export * from './eip4361';
-export * from './external-eip4361';
+export * from './eip4361/eip4361';
+export * from './eip4361/external-eip4361';
