+++ packages/taco/src/conditions/shared.ts
@@ -1,13 +1,11 @@
 import { EthAddressSchema } from '@nucypher/shared';
-import { USER_ADDRESS_PARAM_DEFAULT } from '@nucypher/taco-auth';
+import {
+  USER_ADDRESS_PARAM_DEFAULT,
+  USER_ADDRESS_PARAM_EXTERNAL_EIP4361,
+} from '@nucypher/taco-auth';
 import { z } from 'zod';
 
-import {
-  CONTEXT_PARAM_PREFIX,
-  CONTEXT_PARAM_REGEXP,
-  // TODO consider moving this
-  USER_ADDRESS_PARAM_EXTERNAL_EIP4361,
-} from './const';
+import { CONTEXT_PARAM_PREFIX, CONTEXT_PARAM_REGEXP } from './const';
 
 export const contextParamSchema = z.string().regex(CONTEXT_PARAM_REGEXP);
 // We want to discriminate between ContextParams and plain strings
